requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
motor==3.1.1
multidict==6.0.2
pymongo==4.3.2
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
yarl==1.8.1
